@@ -103,7 +103,7 @@ class BlocProvider<B extends Bloc> extends BlocSubscriber<B, RouteData> {
 }
 
 class _BlocProviderState<B extends Bloc>
-    extends BlocSubscriberState<B, RouteData, BlocProvider<B>> {
+    extends BlocSubscriberState<B, RouteData?, BlocProvider<B>> {
   late final B _bloc;
 
   @override
@@ -122,13 +122,13 @@ class _BlocProviderState<B extends Bloc>
   }
 
   @override
-  Stream<RouteData>? get stream =>
+  Stream<RouteData?>? get stream =>
       widget.router == null ? null : _bloc.navigationStream;
 
   @override
-  void onNewState(RouteData state) {
-    final result = widget.router?.call(context, state.name, state.arguments);
-    state.resultConsumer(result);
+  void onNewState(RouteData? state) {
+    final result = widget.router?.call(context, state?.name, state?.arguments);
+    state?.resultConsumer(result);
   }
 
   @override
